Share ObjectFlag enum
