+++ CSSE/CSBaseObject.h
@@ -2,16 +2,9 @@
 
 #include "CSDefines.h"
 #include "TES3ObjectType.h"
+#include "TES3ObjectFlags.h"
 
 namespace se::cs {
-	namespace ObjectFlag {
-		typedef unsigned int value_type;
-
-		enum Flag : value_type {
-			ScaleModifiedToOne = 0x8000
-		};
-	}
-
 	struct BaseObject_VirtualTable {
 		void(__thiscall* destructor)(BaseObject*, signed char); // 0x0
 		int(__thiscall* loadObjectSpecific)(BaseObject*, GameFile*); // 0x4
