+++ CSSE/CSSE.cpp
@@ -557,12 +557,12 @@ namespace se::cs {
 			CS_Object_SetScale(reference, scale, clamp);
 			const auto newScale = reference->getScale();
 			if (newScale == 1.0f && oldScale != newScale) {
-				reference->flags |= ObjectFlag::ScaleModifiedToOne;
+				reference->flags |= TES3::ObjectFlag::ScaleModifiedToOne;
 			}
 		}
 
 		bool __fastcall shouldSaveReferenceScale(Reference* reference) {
-			return reference->getScale() != 1.0f || (reference->flags & ObjectFlag::ScaleModifiedToOne) != 0;
+			return reference->getScale() != 1.0f || (reference->flags & TES3::ObjectFlag::ScaleModifiedToOne) != 0;
 		}
 
 		__declspec(naked) void PatchSaveReferenceScaleCheck() {
