+++ SharedSE/SharedSE.vcxitems
@@ -117,6 +117,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)LinkedObjectsList.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Deque.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)StdString.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3ObjectFlag.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3ObjectType.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)WindowsUtil.h" />
   </ItemGroup>
