+++ SharedSE/TES3ObjectType.h
